reformat all codes

+++ Assets/Scripts/UI/UnitWorldUI.cs
@@ -1,6 +1,6 @@
-using UnityEngine;
 using TMPro;
 using UnitSystem;
+using UnityEngine;
 using UnityEngine.UI;
 
 namespace UI
@@ -18,33 +18,20 @@ namespace UI
 
         private Transform _cameraTransform;
 
-        private void Awake()
-        {
-            _cameraTransform = Camera.main.transform;
-        }
+        private void Awake() { _cameraTransform = Camera.main.transform; }
 
         private void Start()
         {
             unit.OnActionPointsChanged += UpdateActionPoints;
             unit.Health.OnHealthChanged += UpdateHealth;
-
             UpdateActionPoints();
             UpdateHealth();
         }
 
-        private void Update()
-        {
-            transform.LookAt(_cameraTransform);
-        }
+        private void Update() { transform.LookAt(_cameraTransform); }
 
-        private void UpdateActionPoints()
-        {
-            actionPointsUI.text = unit.ActionPoints.ToString();
-        }
+        private void UpdateActionPoints() { actionPointsUI.text = unit.ActionPoints.ToString(); }
 
-        private void UpdateHealth()
-        {
-            healthBar.fillAmount = (float)unit.Health.Health / (float)unit.Health.MaxHealth;
-        }
+        private void UpdateHealth() { healthBar.fillAmount = unit.Health.Health / (float)unit.Health.MaxHealth; }
     }
 }
